refactor: Extract shared upsertUserSetting helper to deduplicate upsert pattern
The update-then-insert-if-no-rows pattern was duplicated 4 times across
settings sub-modules. Extracted into settings-helpers.js and simplified
all 4 save functions to single-line delegations.
